This isn't my fault
I found the issue with the canvas stuff and it is not the socket.io stuff - that works fine. the problem is drawOnCanvas or its parameter inputs, they havent been normalised (?) the data is not the same across clients but uses the same co-ords and stuff which is why i think it nots drawing at the same time. I tthink.